# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive SaaS backend API built with Go and Gin framework, providing authentication, user management, customer management, email handling, PDF generation, search functionality, and more.",
        "title": "AE SaaS Basic API",
        "termsOfService": "https://ae-saas-basic.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://ae-saas-basic.com/support",
            "email": "support@ae-saas-basic.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/assets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of available static assets",
                "tags": [
                    "static"
                ],
                "summary": "List static assets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset type filter",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/assets/{path}": {
            "get": {
                "description": "Serve static files like CSS, JS, images with appropriate headers",
                "tags": [
                    "static"
                ],
                "summary": "Serve static assets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset path",
                        "name": "path",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asset content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/logo": {
            "get": {
                "description": "Serve the company logo in various formats",
                "tags": [
                    "static"
                ],
                "summary": "Get company logo",
                "parameters": [
                    {
                        "type": "string",
                        "default": "svg",
                        "description": "Logo format (svg|png|jpg)",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logo file"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/templates/{type}/{template}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Serve template files for preview purposes",
                "tags": [
                    "static"
                ],
                "summary": "Preview templates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template type (email|pdf)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Template name",
                        "name": "template",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}